Action types in separate file
